+++ src/Graph.ts
@@ -9,6 +9,7 @@ import Chunk from './Chunk';
 import ExternalModule from './ExternalModule';
 import Module, { defaultAcornOptions } from './Module';
 import {
+	Asset,
 	InputOptions,
 	IsExternal,
 	LoadHook,
@@ -25,10 +26,15 @@ import {
 	WarningHandler,
 	Watcher
 } from './rollup/types';
-import { Asset, createAssetPluginHooks, finaliseAsset } from './utils/assetHooks';
+import { createAssetPluginHooks, EmitAsset, finaliseAsset } from './utils/assetHooks';
 import { load, makeOnwarn, resolveId } from './utils/defaults';
 import ensureArray from './utils/ensureArray';
-import { randomUint8Array, Uint8ArrayToHexString, Uint8ArrayXor } from './utils/entryHashing';
+import {
+	randomUint8Array,
+	Uint8ArrayEqual,
+	Uint8ArrayToHexString,
+	Uint8ArrayXor
+} from './utils/entryHashing';
 import error from './utils/error';
 import first from './utils/first';
 import { isRelative, relative, resolve } from './utils/path';
@@ -62,6 +68,8 @@ export default class Graph {
 	exportShimVariable: GlobalVariable;
 	treeshakingOptions: TreeshakingOptions;
 	varOrConst: 'var' | 'const';
+
+	private createTransformEmitAsset: () => { assets: Asset[]; emitAsset: EmitAsset };
 
 	contextParse: (code: string, acornOptions?: acorn.Options) => Program;
 
@@ -115,6 +123,8 @@ export default class Graph {
 			return this.acornParse(code, { ...defaultAcornOptions, ...options, ...this.acornOptions });
 		};
 
+		const assetPluginHooks = createAssetPluginHooks(this.assetsById);
+
 		this.pluginContext = {
 			watcher,
 			isExternal: undefined,
@@ -128,8 +138,11 @@ export default class Graph {
 				if (typeof err === 'string') throw new Error(err);
 				error(err);
 			},
-			...createAssetPluginHooks(this.assetsById)
+			emitAsset: assetPluginHooks.emitAsset,
+			getAssetFileName: assetPluginHooks.getAssetFileName,
+			setAssetSource: assetPluginHooks.setAssetSource
 		};
+		this.createTransformEmitAsset = assetPluginHooks.createTransformEmitAsset;
 
 		this.resolveId = first(
 			[
@@ -214,8 +227,16 @@ export default class Graph {
 	}
 
 	getCache() {
+		const assetDependencies: string[] = [];
+		this.assetsById.forEach(asset => {
+			if (!asset.transform && asset.dependencies && asset.dependencies.length) {
+				for (const depId of asset.dependencies) assetDependencies.push(depId);
+			}
+		});
+
 		return {
-			modules: this.modules.map(module => module.toJSON())
+			modules: this.modules.map(module => module.toJSON()),
+			assetDependencies
 		};
 	}
 
@@ -541,7 +562,15 @@ export default class Graph {
 			}
 
 			for (const dynamicModule of module.dynamicImportResolutions) {
-				if (dynamicModule.resolution instanceof Module) {
+				if (!(dynamicModule.resolution instanceof Module)) continue;
+				// If the parent module of a dynamic import is to a child module whose graph
+				// colouring is the same as the parent module, then that dynamic import does
+				// not need to be treated as a new entry point as it is in the static graph
+				if (
+					!graphColouring ||
+					(!dynamicModule.resolution.chunkAlias &&
+						!Uint8ArrayEqual(dynamicModule.resolution.entryPointsHash, curEntry.entryPointsHash))
+				) {
 					if (dynamicImports.indexOf(dynamicModule.resolution) === -1) {
 						dynamicImports.push(dynamicModule.resolution);
 						dynamicImportAliases.push(dynamicModule.alias);
@@ -657,14 +686,24 @@ Try defining "${chunkName}" first in the manualChunks definitions of the Rollup 
 						  }
 						: source;
 
-				if (
-					this.cachedModules.has(id) &&
-					this.cachedModules.get(id).originalCode === sourceDescription.code
-				) {
-					return this.cachedModules.get(id);
+				const cachedModule = this.cachedModules.get(id);
+				if (cachedModule && cachedModule.originalCode === sourceDescription.code) {
+					// re-emit transform assets
+					if (cachedModule.transformAssets) {
+						for (const asset of cachedModule.transformAssets) {
+							this.pluginContext.emitAsset(asset.name);
+						}
+					}
+					return cachedModule;
 				}
 
-				return transform(this, sourceDescription, id, this.plugins);
+				return transform(
+					this,
+					sourceDescription,
+					module,
+					this.plugins,
+					this.createTransformEmitAsset
+				);
 			})
 			.then((source: ModuleJSON) => {
 				module.setSource(source);