+++ test/chunking-form/samples/dynamic-import-inline-colouring/_expected/amd/main.js
@@ -1,0 +1,13 @@
+define(['require'], function (require) { 'use strict';
+
+	var foo = "FOO";
+
+	var foo$1 = /*#__PURE__*/Object.freeze({
+		default: foo
+	});
+
+	var main = Promise.resolve().then(function () { return foo$1; });
+
+	return main;
+
+});