+++ test/chunking-form/samples/dynamic-import-inline-colouring/main.js
@@ -1,0 +1,3 @@
+import foo from "./foo.js";
+
+export default import("./foo.js");